Move code from 212 directory to common source set
