--- clion/src/212/main/kotlin/org/rust/clion/profiler/perf/compatUtils.kt
@@ -1,11 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.clion.profiler.perf
-
-import com.intellij.profiler.clion.perf.PerfProfilerSettings
-import com.intellij.profiler.dtrace.SimpleProfilerSettingsState
-
-fun getPerfSettings(): SimpleProfilerSettingsState = PerfProfilerSettings.instance.state
