--- grazie/src/212/main/resources/META-INF/platform-grazie-only.xml
@@ -1,5 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="com.intellij.grazie">
-        <disableChecking language="Rust"/>
-    </extensions>
-</idea-plugin>
