--- src/212/main/resources/META-INF/platform-core.xml
@@ -1,7 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="com.intellij">
-        <!--suppress PluginXmlValidity -->
-        <formattingService implementation="org.rust.ide.formatter.RustfmtFormattingService"/>
-        <postFormatProcessor implementation="org.rust.ide.formatter.processors.RsTrailingCommaFormatProcessor"/>
-    </extensions>
-</idea-plugin>
