+++ src/main/kotlin/org/rust/ide/inspections/import/ui.kt
@@ -15,6 +15,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.impl.FakePsiElement
 import com.intellij.ui.popup.list.ListPopupImpl
 import com.intellij.ui.popup.list.PopupListElementRenderer
+import com.intellij.util.TextWithIcon
 import org.jetbrains.annotations.TestOnly
 import org.rust.cargo.icons.CargoIcons
 import org.rust.cargo.project.workspace.PackageOrigin
@@ -99,7 +100,7 @@ private class ImportCandidatePsiElement(val importCandidate: ImportCandidate) : 
     override fun getParent(): PsiElement? = importCandidate.qualifiedNamedItem.item.parent
 }
 
-abstract class RsImportCandidateCellRendererBase : DefaultPsiElementCellRenderer() {
+private class RsImportCandidateCellRenderer : DefaultPsiElementCellRenderer() {
 
     private val Any.importCandidate: ImportCandidate? get() = (this as? ImportCandidatePsiElement)?.importCandidate
 
@@ -125,12 +126,11 @@ abstract class RsImportCandidateCellRendererBase : DefaultPsiElementCellRenderer
         }
     }
 
-    // BACKCOMPAT: 2021.1. Inline it
-    protected fun textWithIcon(value: Any?): Pair<String, Icon>? {
+    override fun getItemLocation(value: Any?): TextWithIcon? {
         val crate = value?.importCandidate?.qualifiedNamedItem?.containingCrate ?: return null
         return when (crate.origin) {
-            PackageOrigin.STDLIB, PackageOrigin.STDLIB_DEPENDENCY -> crate.normName to RsIcons.RUST
-            PackageOrigin.DEPENDENCY -> crate.normName to CargoIcons.ICON
+            PackageOrigin.STDLIB, PackageOrigin.STDLIB_DEPENDENCY -> TextWithIcon(crate.normName, RsIcons.RUST)
+            PackageOrigin.DEPENDENCY -> TextWithIcon(crate.normName, CargoIcons.ICON)
             else -> null
         }
     }
