+++ src/main/resources/META-INF/rust-core.xml
@@ -3,7 +3,6 @@
 
     <depends>com.intellij.modules.lang</depends>
 
-    <xi:include href="/META-INF/platform-core.xml" xpointer="xpointer(/idea-plugin/*)"/>
     <xi:include href="/META-INF/experiments.xml" xpointer="xpointer(/idea-plugin/*)"/>
 
     <extensions defaultExtensionNs="com.intellij">
@@ -69,10 +68,11 @@
 
         <!-- Formatter -->
 
-        <!--suppress PluginXmlValidity -->
         <lang.formatter language="Rust" implementationClass="org.rust.ide.formatter.RsFormattingModelBuilder"/>
+        <formattingService implementation="org.rust.ide.formatter.RustfmtFormattingService"/>
         <preFormatProcessor implementation="org.rust.ide.formatter.processors.RsMatchArmCommaFormatProcessor"/>
         <preFormatProcessor implementation="org.rust.ide.formatter.processors.RsStatementSemicolonFormatProcessor"/>
+        <postFormatProcessor implementation="org.rust.ide.formatter.processors.RsTrailingCommaFormatProcessor"/>
 
         <!-- Refactoring support -->
